% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca_workflow}
\alias{cca_workflow}
\title{CCA-Based Niche Analysis Workflow Controller}
\usage{
cca_workflow(mode = c("gradient", "group"), ...)
}
\arguments{
\item{mode}{A character string specifying the analysis mode. 
Must be one of \code{"gradient"} or \code{"group"}.}

\item{...}{Additional arguments passed to the specific workflow functions 
(\code{\link{cca_workflow_gradient}} or \code{\link{cca_workflow_group}}).}
}
\value{
A list containing the results of the executed workflow steps. 
  See \code{\link{cca_workflow_gradient}} or \code{\link{cca_workflow_group}} for structure details.
}
\description{
A controller function that orchestrates the CCA-based niche analysis workflow.
It dispatches the execution to either the gradient workflow (Steps 1-4)
or the group workflow (Steps 1-2, 5-6) based on the specified \code{mode}.
}
\examples{
data("plant_otu")
data("plant_env")
otu <- plant_otu[rowSums(plant_otu > 0) >= 15, ]
res <- cca_workflow(
  mode = "gradient",
  otu = otu,
  env = plant_env,
  sel = c("AMP", "Moisture", "AP", "pH"),
  covariates = "AMT",
  var = "AMT",
  method = "loess",
  galaxy_colnum = FALSE,
  make_plot = FALSE,
  top_node = 20
)
str(res, max.level = 1)
}
