% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNiche-package.R
\docType{package}
\name{EcoNiche-package}
\alias{EcoNiche}
\alias{EcoNiche-package}
\title{CoNiche: Community Niche Position and Width Estimation Tools}
\description{
CoNiche provides tools to estimate niche position and niche width using
(i) CCA/partial-CCA workflows, (ii) GAM fits along single gradients, and
(iii) Levins niche breadth metrics, with functions for species- and site-level summaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yedeng-lab/EcoNiche}
  \item Report bugs at \url{https://github.com/yedeng-lab/EcoNiche/issues}
}

}
\author{
\strong{Maintainer}: Shuotao Zhou \email{zhoushuotao@stu.hunau.edu.cn}

Authors:
\itemize{
  \item Kai Feng \email{kaifeng@rcees.ac.cn}
  \item Ye Deng \email{yedeng@rcees.ac.cn}
}

}
\keyword{internal}
