% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_env}
\alias{plant_env}
\title{Example plant environmental table}
\format{
A data frame with 290 rows and 11 variables:
\describe{
  \item{TN}{Total nitrogen.}
  \item{TK}{Total potassium.}
  \item{TP}{Total phosphorus.}
  \item{SOC}{Soil organic carbon.}
  \item{AP}{Available phosphorus.}
  \item{AN}{Available nitrogen.}
  \item{Moisture}{Soil moisture.}
  \item{AMT}{Annual mean temperature.}
  \item{AMP}{Annual mean precipitation.}
  \item{pH}{Soil pH.}
  \item{abspH}{Absolute deviation from neutral pH.}
}
}
\usage{
plant_env
}
\description{
Environmental variables for the samples in \code{\link{plant_otu}}. Rows are
samples and columns are environmental variables.
}
\seealso{
\code{\link{plant_otu}}, \code{\link{plant_group}}
}
\keyword{datasets}
