% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca_workflow}
\alias{cca_workflow}
\title{CCA-Based Niche Analysis Workflow Controller}
\usage{
cca_workflow(mode = c("gradient", "group"), ...)
}
\arguments{
\item{mode}{A character string specifying the analysis mode. 
Must be one of \code{"gradient"} or \code{"group"}.}

\item{...}{Additional arguments passed to the specific workflow functions 
(\code{\link{cca_workflow_gradient}} or \code{\link{cca_workflow_group}}).}
}
\value{
A list containing the results of the executed workflow steps. 
  See \code{\link{cca_workflow_gradient}} or \code{\link{cca_workflow_group}} for structure details.
}
\description{
A controller function that orchestrates the CCA-based niche analysis workflow.
It dispatches the execution to either the gradient workflow (Steps 1-4)
or the group workflow (Steps 1-2, 5-6) based on the specified \code{mode}.
}
\examples{
\dontrun{
# Run the gradient workflow
res <- cca_workflow(mode = "gradient", otu = my_otu, env = my_env, sel = "Temp")
}
}
